+++ CHANGES.rst
@@ -12,7 +12,7 @@ There's a frood who really knows where his towel is.
 2.6b1 (2016-12-21)
 ^^^^^^^^^^^^^^^^^^
 
-- Add Twitter Cards metadata (closes `#65`_).
+- Add metadata for Twitter Cards (closes `#65`_).
   [rodfersou]
 
 - Code clean up; tests related with loading BMP images were removed as make no sense.